feat(Schedule): move code from World to Schedule
